pyproject.toml
[tool.poetry]
name = "avocados-technical-task"
version = "0.1.0"
description = "Technical task about building a simple ML model in Numpy and a Flask API around it."
authors = ["Roberto Fierimonte <roberto.fierimonte@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.23.4"
flask = "^2.2.2"
flask-mysql = "^1.5.2"
loguru = "^0.6.0"
joblib = "^1.2.0"
flask-sqlalchemy = "^3.0.2"
marshmallow = "^3.19.0"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.28.1"
pymysql = "^1.0.2"
mysqlclient = "^2.1.1"
cryptography = "^38.0.3"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
commitizen = "^2.37.0"
black = {version = "^22.10.0", allow-prereleases = true}
ipykernel = "^6.17.0"
pandas = "^1.5.1"
numba = "^0.56.4"
matplotlib = "^3.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
